Massive change - separated to different libs so its more manageable
